--- pkg/onit/setup2/k8s/cli.go
@@ -1,1 +1,0 @@
-package k8s
