--- pkg2/kubetest/env.go
@@ -1,1 +1,0 @@
-package kubetest
